fantasy-hockey-analytics/requirements.txt
yfpy
python-dotenv
